+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add on_select attribute on tab slot
 - Add scroll area component [#296](https://github.com/mishka-group/mishka_chelekom/pull/296) [#300](https://github.com/mishka-group/mishka_chelekom/pull/300)
 - Add missing features of Phoenix core components [#296](https://github.com/mishka-group/mishka_chelekom/pull/296)
+- Add global components installing support and JS formatter [#304](https://github.com/mishka-group/mishka_chelekom/pull/304)
 
 ### Bugs:
 - Add missing slot to gallery component [#294](https://github.com/mishka-group/mishka_chelekom/pull/294)